Acknowledge cljs dependency in `tools.deps` docs
When I followed the instructions to run an nREPL using clj I got a
compilation error containing the following message:

  java.io.FileNotFoundException: Could not locate
  clojure/tools/reader__init.class, clojure/tools/reader.clj or
  clojure/tools/reader.cljc on classpath.

Adding a deps.edn dependency on clojure.tools.reader fixes that, but
reveals another error. It eventually becomes clear that Piggieback
simply depends on ClojureScript.

This commit adds a note to the README to help the next person.