README: Add notes about starting CLJS REPL
When I tried to set up a CLJS REPL in Cursive for the first time, I only
followed the instructions in section ‘Installation’. It didn't work.
Then I went only through a variation of the steps in section ‘Usage’. It
didn't work. Then I became hysterical and tried anything and looked
everywhere I could think of. Finally I realized that I needed to start
an nREPL, then connect to it, then start the CLJS REPL within it.

Add a few sentences that will hopefully save other people a similar
odyssey.