Implement nREPL's pretty printing functionality (#108)
Try to read the ClojureScript evaluation result in Clojure and rely on the nREPL
printing middleware to print it if reading succeeds. The type
`UnknownTaggedLiteral` is introduced to be able to read and print tagged
literals which are specific to ClojureScript such as `#queue` (or any
user-defined ones). The output may vary based on the printing function
used (for example fipp and puget don't rely on the supplied
print-method for `UnknownTaggedLiteral`).

Reading can still fail in some cases, so a fallback is also included
by trying to detect if pretty printing is enabled
(e.g. `nrepl.middleware.print/print` not set to `cider.nrepl.pprint/pr` for tools using `cider-nrepl`) and
if that is the case, wrapping the ClojureScript form with `cljs.pprint/pprint`
(instead of the default `cljs.core.pr-str`). This way pretty printing
can still be turned on or off, even when reading fails.

Co-authored-by: Shen Tian <biozinc@gmail.com>